requirements.txt
tritonclient[grpc]==2.36.0
tensorflow==2.13.1
tensorflow-serving-api==2.13.1
numpy==1.24.3
tqdm==4.66.0
